add message label and refactor dto
