+++ src/main/java/ru/startup/game/chatserver/model/dto/MessageLabelDto.java
@@ -1,0 +1,14 @@
+package ru.startup.game.chatserver.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class MessageLabelDto {
+
+    private Long messageId;
+    private String labelName;
+}
